requirements.txt
# Core Dependencies
diskcache==5.6.3
ipdb==0.13.13
langchain_openai==0.0.8
numpy==1.26.4
openai_async==0.0.3
Pillow==10.2.0
pyautogen==0.2.15
pymilvus==2.3.4
PyYAML==6.0.1
redis==5.0.2
setuptools==65.5.0
SQLAlchemy==1.4.22
starlette==0.36.3
typing_extensions==4.10.0
uvicorn==0.28.0
pymysql==1.1.0
scikit-learn==1.4.0
dill
fastapi==0.110.0
pydantic==2.6.4

# Documentation Dependencies
mkdocs==1.6.1
mkdocs-material==9.5.49
mkdocs-material-extensions==1.3.1
mkdocs-mermaid2-plugin==1.2.1
pymdown-extensions==10.13
pygments==2.19.1
pillow==10.2.0
cairosvg==2.7.1  # For PDF export support
mike==2.0.0  # For documentation versioning
mkdocs-minify-plugin==0.8.0  # For optimizing build size
mkdocs-git-revision-date-plugin==0.3.2  # For showing last update dates
mkdocs-redirects==1.2.1  # For handling redirects
mkdocs-awesome-pages-plugin==2.9.2  # For better navigation control

# Development & Testing
pytest==8.0.0
pytest-cov==4.1.0
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0